Data Structures

Work in progress...

Main.OCTHaGOn.GlobalModel — Type

Contains all required info to be able to generate a global optimization problem. NOTE: proper construction is to use addnonlinearconstraint to add bbls. model must be a mixed integer convex model. nonlinear_model can contain JuMP.NonlinearConstraints.

Main.OCTHaGOn.BlackBoxLearner — Type

BBL type is for function definitions!

Main.OCTHaGOn.BlackBoxClassifier — Type

@with_kw mutable struct BlackBoxClassifier

Allows for approximation of constraints using OCTs. To be added to GlobalModel.bbls using functions: addnonlinearconstraints addnonlinearor_compatible

Mandatory arguments are: vars::Array{JuMP.VariableRef,1}

Other arguments may be necessary for proper functioning: For data-driven constraints, need: X::DataFrame Y:: Array For constraint functions, need : constraint::Union{JuMP.ConstraintRef, Expr}

Optional arguments: exprvars::Union{Array, Nothing} JuMP variables as function arguments (i.e. vars rolled up into vector forms). vars ⋐ flat(exprvars) name::String equality::Bool Specifies whether function should be satisfied to an equality

Main.OCTHaGOn.BlackBoxRegressor — Type

@with_kw mutable struct BlackBoxRegressor

Allows for approximation of constraints using OCTs. To be added to GlobalModel.bbls using functions: addnonlinearconstraints addnonlinearor_compatible

Mandatory arguments are:
    vars::Array{JuMP.VariableRef,1}
    dependent_var::JuMP.VariableRef

Other arguments may be necessary for proper functioning: For data-driven constraints, need: X::DataFrame Y:: Array For constraint functions, need : constraint::Union{JuMP.ConstraintRef, Expr}

Optional arguments: exprvars::Union{Array, Nothing} JuMP variables as function arguments (i.e. vars rolled up into vector forms). vars ⋐ flat(exprvars) name::String equality::Bool Specifies whether function should be satisfied to an equality

Main.OCTHaGOn.LinkedLearner — Type

Superclass of LinkedClassifier and LinkedRegressor.

Main.OCTHaGOn.LinkedClassifier — Type

Contains data for a constraint that is repeated.

Main.OCTHaGOn.LinkedRegressor — Type

Contains data for a constraint that is repeated.