Sampling methods

Work in progress...

Main.OCTHaGOn.uniform_sample_and_eval! — Function

uniform_sample_and_eval!(bbl::Union{BlackBoxClassifier, BlackBoxRegressor}; boundary_fraction, lh_iterations, sample_density)

Samples and evaluates nonlinear constraints using full suite of methods.

Main.OCTHaGOn.lh_sample — Function

lh_sample(vars::Array{JuMP.VariableRef, 1}; lh_iterations::Int64 = 0,
               n_samples::Int64 = 1000)
lh_sample(bbl::BlackBoxLearner; lh_iterations::Int64 = 0,
               n_samples::Int64 = 1000)

Uniformly Latin Hypercube samples the variables of GlobalModel, as long as all lbs and ubs are defined.

Main.OCTHaGOn.boundary_sample — Function

boundary_sample(bbl::BlackBoxLearner; fraction::Float64 = 0.5)
boundary_sample(vars::Array{JuMP.VariableRef, 1}; n_samples = 100, fraction::Float64 = 0.5,
                     warn_string::String = "")

Samples a BlackBoxLearner on the corners of the variable hypercube. Samples a subset of corners for learners with large number of variables.

Main.OCTHaGOn.knn_sample — Function

knn_sample(bbl::BlackBoxClassifier; k::Int64 = 10, sample_density = 1e-5, sample_idxs = nothing)

Implements KNN sampling on a BlackBoxLearner. Note: must have at least one in/feasible point for the algorithm.