Data structures

Work in progress...

Main.OCTHaGOn.GlobalModel — Type

Contains all constraints and data to solve the global optimization optimization problem. Add linear constraints to a GlobalModel through JuMP.@constraint(gm.model, ...), and nonlinear constraints using add_nonlinear_constraint.

Main.OCTHaGOn.BlackBoxLearner — Type

Union of BlackBoxClassifier and BlackBoxRegressor types.

Main.OCTHaGOn.BlackBoxClassifier — Type

Allows for approximation of constraints using OCTs. Can be added to GlobalModel using add_nonlinear_constraint or add_nonlinear_or_compatible.

Main.OCTHaGOn.BlackBoxRegressor — Type

Allows for approximation of continuous functions or objectives using ORTs. Can be added to GlobalModel using add_nonlinear_constraint or add_nonlinear_or_compatible.

Main.OCTHaGOn.LinkedLearner — Type

Union of LinkedClassifier and LinkedRegressor.

Main.OCTHaGOn.LinkedClassifier — Type

Contains data for a constraint that is repeated. Add to GlobalModel using add_linked_constraint.

Main.OCTHaGOn.LinkedRegressor — Type

Contains data for a constraint that is repeated. Add to GlobalModel using add_linked_constraint.