Sampling methods

Work in progress...

Main.OCTHaGOn.uniform_sample_and_eval! — Function

uniform_sample_and_eval!(bbl::Union{BlackBoxLearner, GlobalModel, Array{BlackBoxLearner}};
                          boundary_fraction::Float64 = 0.5,
                          lh_iterations::Int64 = 0)

Uniform samples and evaluates a BlackBoxLearner. Furthermore, sets the big-M value. Keyword arguments: boundaryfraction: maximum ratio of boundary samples lhiterations: number of GA populations for LHC sampling (0 is a random LH.)

Main.OCTHaGOn.lh_sample — Function

lh_sample(vars::Array{JuMP.VariableRef, 1}; lh_iterations::Int64 = 0,
               n_samples::Int64 = 1000)
lh_sample(bbl::BlackBoxLearner; lh_iterations::Int64 = 0,
               n_samples::Int64 = 1000)

Uniformly Latin Hypercube samples the variables of GlobalModel, as long as all lbs and ubs are defined.

Main.OCTHaGOn.boundary_sample — Function

boundary_sample(bbl::BlackBoxLearner; fraction::Float64 = 0.5)
boundary_sample(vars::Array{JuMP.VariableRef, 1}; n_samples = 100, fraction::Float64 = 0.5,
                     warn_string::String = "")

Smartly samples the constraint along the variable boundaries. NOTE: Because we are sampling symmetrically for lower and upper bounds, the choose coefficient has to be less than ceil(half of number of dims).

Main.OCTHaGOn.knn_sample — Function

knn_sample(bbl::BlackBoxClassifier; k::Int64 = 10, sample_density = 1e-5, sample_idxs = nothing)

Does KNN and secant method based sampling once there is at least one feasible sample to a BlackBoxLearner.